+++ core/src/main/scala/better/files/Dispose.scala
@@ -96,6 +96,10 @@ class Dispose[A](private[Dispose] val resource: A)(implicit disposer: Disposable
     this
   }
 
+  /** Generate a self closing iterator from this disposable resource */
+  def iterator[B](f: A => Iterator[B]): Iterator[B] =
+    CloseableIterator(f(resource), disposeOnce)
+
   /** This keeps the resource open during the context of this flatMap and closes when done */
   def flatMap[B, F[_]](f: A => F[B])(implicit fv: Dispose.FlatMap[F]): fv.Output[B] =
     fv.apply(this)(f)