+++ CHANGES.md
@@ -16,7 +16,7 @@ better-files follows the following `MAJOR.MINOR.PATCH` release conventions:
 ## v4.0.0 [WIP]
 * [x] [PR #584](https://github.com/pathikrit/better-files/pull/584): Remove implicit options from all APIs.
   - Migration Guide: Some API calls would need extra parens e.g. `file.size` is now `file.size()`
-* [ ] [PR #587](https://github.com/pathikrit/better-files/pull/587): `CloseableIterators` fixes resource leak on partial iterations (#295, #403)
+* [x] [Issue #295](https://github.com/pathikrit/better-files/pull/587): `CloseableIterators` fixes resource leaks on partial iterations ([#270](https://github.com/pathikrit/better-files/issues/270), [#403](https://github.com/pathikrit/better-files/issues/403), [#553](https://github.com/pathikrit/better-files/issues/553))
 * [ ] [Issue #129](https://github.com/pathikrit/better-files/issues/129): JSR-203 and JimFS compatibility
 * [ ] [Issue #88](https://github.com/pathikrit/better-files/issues/88): Strongly typed relative and absolute path APIs
 * [ ] [Issue #204](https://github.com/pathikrit/better-files/issues/204): Universal converter APIs